+++ lib/bright_serializer/entity/base.rb
@@ -16,6 +16,7 @@ module BrightSerializer
       def to_h
         @definition = Inflector.deep_transform_keys_in_object(@definition) { |k| Inflector.camel_lower k.to_s }
         parse_ref!
+        evaluate_callable!
         @definition
       end
 
@@ -26,6 +27,12 @@ module BrightSerializer
         ref_entity_name = Inflector.constantize(object.delete('ref')).entity_name
         relation = "#/definitions/#{ref_entity_name}"
         object['$ref'] = relation
+      end
+
+      def evaluate_callable!
+        @definition = Inflector.deep_transform_values_in_object(@definition) do |value|
+          value.respond_to?(:call) ? value.call : value
+        end
       end
 
       def nested_hash(obj, key)