+++ spec/lib/bright_serializer/entity/base_spec.rb
@@ -30,6 +30,14 @@ RSpec.describe BrightSerializer::Entity::Base do
         expect(subject).to eq(['type' => :string])
       end
     end
+
+    context 'when passing a callable object as value' do
+      let(:instance) { described_class.new(type: :string, enum: -> { [1, 2] }) }
+
+      it 'return the definition' do
+        expect(subject).to eq('type' => :string, 'enum' => [1, 2])
+      end
+    end
   end
 
   describe '#parser_ref!' do